+++ Changelog.md
@@ -55,6 +55,7 @@
 - BC: The unused `$config` parameter has been removed from the `RequestStore` checkout driver constructor
 - BC: Added the `deleteByType()` and `clear()` methods to the `AdjustmentCollection` interface
 - BC: Changed the `TaxRate` interface so that it extends the `Configurable` interface
+- BC: Changed the ShippingFeeCalculator, Carrier and Shipment interfaces to Configurable & Schematized
 - BC: Added the `getConfigurationSchema()` method to the `Configurable` interface
 - Added the `Schematized` interface
 - Added the nette/schema package requirement (v1.2.5+)